dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_readfile_event.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

"use strict";

let Ci = SpecialPowers.Ci;
SimpleTest.waitForExplicitFinish();

function createRequest(aType, aOptions, aCallback) {
  let vfsService = SpecialPowers.Cc["@mozilla.org/tv/fakevirtualfilesystemservice;1"].getService(Ci.nsIVirtualFileSystemService);
  let requestManager = vfsService.getRequestManagerById(aOptions.fileSystemId);
  return requestManager.createRequest(aType, aOptions, aCallback);
}

function testEvent() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  return new Promise(function(resolve, reject) {
    let _fileSystemId = "dummyId";
    let _displayName = "dummyFileSystem";
    let _writable = true;
    let _openedFilesLimit = 10;
    let _requestId = 0;
    let _openRequestId = 1;
    let _offset = 0;
    let _length = 5;
    let successCallback = {
      onSuccess: function(requestId, value, hasMore) {
        ok(true, "successCallback is invoked.");
        let val = value.QueryInterface(Ci.nsIVirtualFileSystemReadFileRequestValue);
        is(hasMore, false, "hasMore must be false");
        is(val.data, "ABCDE", "data is correct");
        resolve();
      },
      onError: function(requestId, error) {}
    };
    navigator.fileSystemProvider.addEventListener("readfilerequested", function onReadFileRequested(event){
      navigator.fileSystemProvider.removeEventListener("readfilerequested", onReadFileRequested);
      ok(true, "got event.");
      is(event.options.fileSystemId, _fileSystemId, "file system id must be the same");
      is(event.options.openRequestId, _openRequestId, "open request id must be the same");
      is(event.options.offset, _offset, "offset must be the same");
      is(event.options.length, _length, "length must be the same");
      var buf = new ArrayBuffer(3);
      var view = new Uint8Array(buf);
      // set the content to "ABC"
      view[0] = 65;
      view[1] = 66;
      view[2] = 67;
      event.successCallback(buf, true);
      var buf1 = new ArrayBuffer(2);
      view = new Uint8Array(buf1);
      // set the content to "ABC"
      view[0] = 68;
      view[1] = 69;
      event.successCallback(buf1, false);
    });
    navigator.fileSystemProvider.mount({fileSystemId: _fileSystemId,
                              displayName: _displayName,
                              writable: _writable,
                              openedFilesLimit: _openedFilesLimit}).then(
      function() {
        ok(true, "mount ok.");

        let options = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-readfile-requested-options;1"].createInstance(Ci.nsIVirtualFileSystemReadFileRequestedOptions);
        options.fileSystemId = _fileSystemId;
        options.openRequestId = _openRequestId;
        options.offset = _offset;
        options.length = _length;
        _requestId = createRequest(Ci.nsIVirtualFileSystemRequestManager.REQUEST_READFILE,
                                   options,
                                   successCallback);
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        resolve();
      }
    );
  });
}


function runTest() {
  ok(navigator.fileSystemProvider, "should have navigator.fileSystemProvider");

  testEvent()
  .then(function() {
    info("test finished");
    SimpleTest.finish();
  });
}

SpecialPowers.pushPrefEnv({"set": [["device.storage.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{"type": "filesystemprovider", "allow": true, "context": document}], runTest);
});

 </script>
 </pre>
 </body>
 </html>
